refactor(screener): extract URL/form sync and query normalizer; refactor ScreenerPage for readability (refs #6)
