+++ src/features/tickers/query.ts
@@ -1,0 +1,8 @@
+import type { TickerRow } from '../../lib/types';
+
+export type TickerQueryData = TickerRow[] | { rows: TickerRow[] } | undefined;
+
+export function toTickerRows(input: TickerQueryData): TickerRow[] {
+  if (!input) return [];
+  return Array.isArray(input) ? input : input.rows;
+}
